register working

+++ src/main/java/com/taxApp/service/AppUserService.java
@@ -3,7 +3,7 @@ package com.taxApp.service;
 import com.taxApp.model.AppUser;
 
 public interface AppUserService {
-	public int insertAppUser(AppUser _user);
+	public long insertAppUser(AppUser _user);
 	public AppUser findUserByEmail(String email);
 	public AppUser findUserById(int id);
 }
